tests: move user subcommand tests
Move the unit tests relating to the association_table
in FluxAccounting.db to their own test file
